Refactor config and main:
Change the name of `set() to `server_config()`.

Move `servers` and `start` to `lib.rs`
